+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-0.6.0 (The future)
+0.6.0 (2023-04-04)
 ==================
 ### New Features
 - Variance of the weighted mean
@@ -10,6 +10,8 @@
         Computes the confidence intervals of the weighted mean using the (just added) variance of the weighted mean.
         - Added the `ci_of_mean` option to stat in the `descriptive_stats` function (based on `ci_of_weighted_mean`). Also added kwargs support.
         - Added the `.ci_of_mean()` method to BalanceDF.
+        - Added the `.mean_with_ci()` method to BalanceDF.
+        - Updated `.summary()` methods to include the output of `ci_of_mean`.
 - Improve 'choose_variables' function to control the order of the returned variables
     - The return type is now a list (and not a Tuple)
     - The order of the returned list is based on the variables argument. If it is not supplied, it is based on the order of the column names in the DataFrames. The df_for_var_order arg controls which df to use.